# [MISC] Improve bulk_contains performance 