--- a/test/performance/ibf/CMakeLists.txt
+++ b/test/performance/ibf/CMakeLists.txt
@@ -1,0 +1,1 @@
+hibf_benchmark (interleaved_bloom_filter_benchmark.cpp)