Things that will also help:
   * aligned alloc

Things that might help:
  * range interface for `bulk_contains`, especially in conjunction with `bulk_count`

The changes rely on auto-vectorization and might be done more elegantly by hand.

`192` is faster than before, but relatively slow. `128` and `256` can be done in one step with AVX2, but `192` requires two. `192` is around 30% slower than `128` or `256`.

bulk_contains benchmark with 1GiB data, throughput in hashes/sec:

| bins | old   | new    | speedup |
|------|-------|--------|---------|
| 64   | 78.30 | 117.13 | 1.50    |
| 128  | 59.33 | 121.20 | 2.04    |
| 192  | 41.43 | 84.47  | 2.04    |
| 256  | 42.04 | 122.27 | 2.91    |
| 1024 | 16.50 | 95.74  | 5.80    |